requirements.txt
bcrypt==4.0.1
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0
python-dateutil==2.8.2
yt-dlp==2023.3.4
fastapi==0.68.1
uvicorn==0.15.0 